+++ PGPatcher/include/GUI/ModSortDialog.hpp
@@ -11,6 +11,7 @@
 #include <wx/sizer.h>
 #include <wx/wx.h>
 
+#include <set>
 #include <string>
 #include <unordered_map>