+++ PGPatcher/src/PGPatcherGlobals.cpp
@@ -1,5 +1,4 @@
 #include "PGPatcherGlobals.hpp"
-#include "PGGlobals.hpp"
 
 ParallaxGenConfig* PGPatcherGlobals::s_PGC = nullptr;
 auto PGPatcherGlobals::getPGC() -> ParallaxGenConfig* { return s_PGC; }